+++ js/cmd.js
@@ -1,10 +1,14 @@
 #!/usr/bin/env node
 'use strict';
 
-const program = require('commander');
+const commander = require('commander'); // (normal include)
+const program = new commander.Command();
 const execute = require('./index');
+const compare = require('./compare');
 
 program
+  .command('scan', { isDefault: true })
+  .description('scan a list of urls to gather Google lighthouse scores')
   .option(
     '-s, --sites [sites]',
     'a comma delimited list of site urls to analyze with Lighthouse'
@@ -27,6 +31,22 @@ program
   .option('-m, --mobile-only', 'Only run using mobile profile')
   .option('-q, --quiet', 'Hide all console output')
   .version(require('../package.json').version)
-  .parse(process.argv);
+  .action((options) => {
+    execute(options);
+  });
 
-execute(program.opts());
+
+program
+  .command('compare')
+  .description('compare sites scans over time to monitor progress')
+  .option(
+    '-f, --folder <path>',
+    'name of a absolute path, relative path, or name of a directory within /lighthouse-mariner-reports folder'
+  )
+  .version(require('../package.json').version)
+  .action((options) => {
+    compare(options);
+  });
+
+program.parse(process.argv);
+