+++ js/index.js
@@ -5,19 +5,27 @@ const chalk = require('chalk');
 const Utils = require('./utils');
 const { logger } = Utils;
 const Output = require('./output');
+const _ = require('lodash');
 
-let report = [];
+let report = {
+  info: {
+    date: new Date().getTime(),
+    scales: {},
+  },
+  scans: [],
+};
 
 module.exports = execute;
 
 async function execute(options) {
   var hrstart = process.hrtime();
 
-  setup(options);
-
   const devices = Utils.getDevices(options);
   const passes = options.passes;
   const sites = getSites(options);
+  const scanDomain = sites[0].name;
+
+  setup(options, scanDomain);
 
   logger.log(chalk.magentaBright(`\nURLs to be scanned: ${sites.length}`));
   logger.log(chalk.magenta(`\Passes on each URL: ${passes}`));
@@ -34,7 +42,13 @@ async function execute(options) {
         Utils.deviceConfigs[device].flags,
         Utils.deviceConfigs[device].config
       );
-      report.push(result);
+
+      report.info.scales = result.scales;
+      report.info.version = result.version;
+      delete result.scales;
+      delete result.version;
+
+      report.scans.push(result);
       Output.jsonReport(report);
     }
   }
@@ -47,7 +61,7 @@ async function execute(options) {
   logger.log(`Execution time: ${perfTime[0]}s ${perfTime[1] / 1000000}ms`);
 }
 
-function setup(opts) {
+function setup(opts, scanDomain) {
   if (!opts.file && !opts.sites) {
     console.error('No URLs or file was specified');
     return;
@@ -56,6 +70,8 @@ function setup(opts) {
   if (opts.quiet) {
     logger.quietLogs();
   }
+
+  // opts.outputDirectory = opts.outputDirectory + '/' + scanDomain;
 
   Output.setOutputOptions(opts);
 }
@@ -100,35 +116,33 @@ function getSites(options) {
 
 function summarize(report) {
   let summary = {
-    scales: report[0].scales,
+    ...report.info,
     scores: [],
-    version: report[0].version,
   };
 
-  report.forEach((item) => {
+  report.scans.forEach((item) => {
     summary.scores[item.url] = summary.scores[item.url] || {};
 
     // Scores average calculations
     for (let x in item.scores) {
-      summary.scores[item.url][x] = summary.scores[item.url][x] || {};
-      summary.scores[item.url][x][item.device] = Math.round(
-        Utils.average(item.scores[x])
+      _.setWith(
+        summary.scores[item.url],
+        `${x}.${item.device}`,
+        Math.round(Utils.average(item.scores[x])),
+        Object
       );
     }
 
     // Metrics average calculations
     for (let x in item.metrics) {
-      summary.scores[item.url]['metrics'] =
-        summary.scores[item.url]['metrics'] || {};
-      summary.scores[item.url]['metrics'][x] =
-        summary.scores[item.url]['metrics'][x] || {};
-      summary.scores[item.url]['metrics'][x][item.device] =
-        summary.scores[item.url]['metrics'][x][item.device] || {};
-      summary.scores[item.url]['metrics'][x][item.device].value = Utils.average(
-        item.metrics[x].value
-      );
-      summary.scores[item.url]['metrics'][x][item.device].score = Math.round(
-        Utils.average(item.metrics[x].score)
+      _.setWith(
+        summary.scores[item.url],
+        `metrics.${x}.${item.device}`,
+        {
+          value: Utils.average(item.metrics[x].value),
+          score: Math.round(Utils.average(item.metrics[x].score)),
+        },
+        Object
       );
     }
   });